+++ src/vue/pages/evm-explorer/tabs/evm-explorer-overview/EvmBlockRow.vue
@@ -19,7 +19,7 @@
       {{ block.minerHash }}
     </router-link>
     <p>
-      {{ block.totalTransactionCount }}
+      {{ $fnumber(block.totalTransactionCount) }}
     </p>
     <p :key="timeNow">
       {{ timeAgo() }}